Fixed vessel movement interpolation

+++ Common/Message/Data/Vessel/VesselUpdateMsgData.cs
@@ -42,6 +42,6 @@ namespace LunaCommon.Message.Data.Vessel
         public double[] Velocity { get; set; } //Only if IsSurfaceUpdate
         public double[] Acceleration { get; set; } //Only if IsSurfaceUpdate
         public double[] Orbit { get; set; } //Only if NOT IsSurfaceUpdate
-        public long SentTime { get; set; }
+        public float SentTime { get; set; }
     }
 }
